+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
-All notable changes to the "reactjs-vscode-helper" extension will be documented in this file.
+<!-- All notable changes to the "reactjs-vscode-helper" extension will be documented in this file. -->
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+<!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
-## [Unreleased]
+## [v0.1.0] - 2023-12-03
 
 - Initial release
+
+### Added
+- Can convert react default imports to React.lazy imports